Got turns with PD controller
I got it to do two turns with the PD controller, but overall it wasn't really smooth at all. Then the next go I tried again and my batteries died so I put them on charge. I should have recorded it, but I wasn't thinking that it would actually work. So I don't know if it was low batteries that allowed it to make the turns or the actual code is working properly.