Changed up code... Battery died again...
So I juiced the batteries a little and tried again. I made some changes. So I think if you just keep tweaking the Kp, Ki and Kd values we should get smoother turns and stuff. I just can't test it anymore and I tried to record it, but I forgot my phones camera is broken. So the recording is shaking all over the place.